refactor: split configuration page into tabbed sub-pages
Split the monolithic /configuration page into four independent
route-based tabs: Application, Backup, Notification, and Authentication.
Each tab is its own Livewire component with dedicated route, improving
maintainability and navigation.
