+++ resources/views/layouts/app.blade.php
@@ -42,7 +42,7 @@
                 <x-menu-item title="{{ __('Users') }}" icon="o-users" link="{{ route('users.index') }}" wire:navigate />
                 <x-menu-item title="{{ __('Agents') }}" icon="o-cpu-chip" link="{{ route('agents.index') }}" wire:navigate :badge="__('Beta')" badge-classes="badge-warning badge-soft badge-xs" />
                 <x-menu-separator />
-                <x-menu-item title="{{ __('Configuration') }}" icon="o-cog-6-tooth" link="{{ route('configuration.index') }}" wire:navigate />
+                <x-menu-item title="{{ __('Configuration') }}" icon="o-cog-6-tooth" link="{{ route('configuration.application') }}" wire:navigate />
                 <x-menu-item title="{{ __('API Docs') }}" no-wire-navigate="true" icon="o-document-text" link="{{ route('scramble.docs.ui') }}" />
                 <x-menu-item title="{{ __('API Tokens') }}" icon="o-key" link="{{ route('api-tokens.index') }}" wire:navigate />
             </x-menu>
